Split tests into e2e, integration and unit
